requirements.txt
# Core Dependencies
        Flask==3.1.0
        flask-jwt-extended==4.4.4
        Flask-Cors==3.0.10
        Flask-Session==0.8.0
        Flask-Talisman==1.0.0
        #Flask-Login==0.6.3
        python-dotenv==1.0.0
        bcrypt==4.0.1
        tensorflow==2.16.2
        nltk==3.9.0
        scikit-learn==1.3.1
        numpy==1.26.4
        pandas==2.1.1
        matplotlib==3.8.0
        mysql-connector-python==8.1.0
        cryptography==42.0.5
        pycryptodome==3.21.0
        setuptools==69.1.1
        wheel==0.42.0
        json5==0.9.6
        gunicorn==20.1.0

        # Testing Frameworks
        pytest==7.4.0
        pytest-cov==4.1.0

        # End-to-End and Usability Testing
        selenium==4.12.0

        # Load Testing
        locust==2.15.1

        # Security Testing
        python-owasp-zap-v2.4==0.0.14